Update __main__.py
added 'hevc_nvenc' as the default encoder, this uses the Nvidia GPU for hardware encoding.
Set '--ffmpeg-location' to 'ffmpeg', which uses the ffmpeg binary in the folder and if no binary is in the folder it uses ffmpeg in your path. If there is no ffmpeg in either of these places it will error out.